chore: isolate segment matching example
