+++ includes/class-newspack-popups-criteria.php
@@ -163,8 +163,7 @@ final class Newspack_Popups_Criteria {
 			self::SCRIPT_HANDLE,
 			'newspackPopupsCriteria',
 			[
-				'config'   => self::get_criteria_config(),
-				'segments' => Newspack_Popups_Segmentation::get_segments( false ),
+				'config' => self::get_criteria_config(),
 			]
 		);
 		wp_enqueue_script(
@@ -175,6 +174,14 @@ final class Newspack_Popups_Criteria {
 			true
 		);
 		wp_script_add_data( 'newspack-popups-default-criteria', 'defer', true );
+		wp_enqueue_script(
+			'newspack-popups-segments-example',
+			plugins_url( '../dist/segmentsExample.js', __FILE__ ),
+			[ self::SCRIPT_HANDLE ],
+			filemtime( dirname( NEWSPACK_POPUPS_PLUGIN_FILE ) . '/dist/segmentsExample.js' ),
+			true
+		);
+		wp_script_add_data( 'newspack-popups-segments-example', 'defer', true );
 	}
 
 	/**
