+++ src/criteria/index.js
@@ -96,7 +96,7 @@ function registerCriteria( config ) {
 		// Set criteria value.
 		const ras = window.newspackReaderActivation;
 		if ( ! ras ) {
-			console.warn( `Reader activation script has not loaded.` ); // eslint-disable-line no-console
+			console.warn( 'Reader activation script not loaded.' ); // eslint-disable-line no-console
 		}
 		if ( typeof criteria.matchingAttribute === 'function' ) {
 			criteria.value = criteria.matchingAttribute( ras );
@@ -115,46 +115,3 @@ function registerCriteria( config ) {
 if ( newspackPopupsCriteria?.config?.length ) {
 	newspackPopupsCriteria.config.forEach( registerCriteria );
 }
-
-window.newspackRAS = window.newspackRAS || [];
-window.newspackRAS.push( () => {
-	/**
-	 * Whether the reader matches the segment criteria.
-	 */
-	const matchSegment = segment => {
-		for ( const criteriaId in segment.criteria ) {
-			const criteria = registeredCriteria[ criteriaId ];
-			if ( ! criteria ) {
-				continue;
-			}
-			const config = segment.criteria[ criteriaId ];
-			if ( ! criteria.matches( config ) ) {
-				return false;
-			}
-		}
-		return true;
-	};
-	newspackPopupsCriteria.matchSegment = matchSegment;
-
-	/**
-	 * Sample segments.
-	 */
-	const segments = [
-		{
-			id: 'segment-1',
-			criteria: {
-				articles_read: { min: 10 },
-			},
-		},
-	];
-	/**
-	 * Match sample segments.
-	 */
-	for ( const segment of segments ) {
-		// eslint-disable-next-line no-console
-		console.log( {
-			segmentId: segment.id,
-			matched: matchSegment( segment ),
-		} );
-	}
-} );
