+++ webpack.config.js
@@ -22,6 +22,7 @@ const blocks = path.join( __dirname, 'src', 'blocks' );
 const customizer = path.join( __dirname, 'src', 'customizer' );
 const criteria = path.join( __dirname, 'src', 'criteria' );
 const defaultCriteria = path.join( __dirname, 'src', 'criteria', 'default' );
+const segmentsExample = path.join( __dirname, 'src', 'criteria', 'segments-example' );
 
 const webpackConfig = getBaseWebpackConfig(
 	{ WP: true },
@@ -36,6 +37,7 @@ const webpackConfig = getBaseWebpackConfig(
 			customizer,
 			criteria,
 			defaultCriteria,
+			segmentsExample,
 		},
 		'output-path': path.join( __dirname, 'dist' ),
 	}
